+++ org.TEC.Proyecto#1/src/Json/Eliminar.java
@@ -5,26 +5,26 @@ import javax.swing.JOptionPane;
 
 public class Eliminar {
 
-    public static void EliminarArchivo(String NombreDeArchivo) {
+    public static void Archivo(String NombreDeArchivo) {
 
         File archivo = new File(RutaBase.RutaBase() + NombreDeArchivo + ".txt");
 
-        if (!archivo.exists()) {
-            JOptionPane.showMessageDialog(null, "No encontrado", "Error", JOptionPane.WARNING_MESSAGE);
-        } else {
+        if (archivo.exists()) {
             archivo.delete();
-            JOptionPane.showMessageDialog(null, "Archivo eliminado de forma exitosa", null, JOptionPane.INFORMATION_MESSAGE);
+            JOptionPane.showMessageDialog(null, "Archivo eliminado de forma exitosa", null, JOptionPane.INFORMATION_MESSAGE);////////////////////////
+        } else {
+            JOptionPane.showMessageDialog(null, "No encontrado", "Error", JOptionPane.WARNING_MESSAGE);                      ////////////////////////
         }
     }
 
-    public static void EliminarTodos(String NombreDeCarpeta) {
+    public static void Todos(String NombreDeCarpeta) {
 
         File Carpeta = new File(RutaBase.RutaBase() + NombreDeCarpeta);
 
-        if (!Carpeta.exists()) {
-            JOptionPane.showMessageDialog(null, "Carpeta no encontrada", "Error", JOptionPane.WARNING_MESSAGE);
-        } else {
+        if (Carpeta.exists() && Carpeta.isDirectory()) {
             VaciarCarpeta(Carpeta);
+        } else {
+            JOptionPane.showMessageDialog(null, "Carpeta no encontrada", "Error", JOptionPane.WARNING_MESSAGE);
         }
     }
 
@@ -38,15 +38,16 @@ public class Eliminar {
         }
     }
 
-    public static void EliminarCarpeta(String NombreDeCarpeta) {
+    public static void Carpeta(String NombreDeCarpeta) {
 
         File Carpeta = new File(RutaBase.RutaBase() + NombreDeCarpeta);
-        VaciarCarpeta(Carpeta);
-        if (!Carpeta.exists()) {
-            JOptionPane.showMessageDialog(null, "Carpeta no encontrada", "Error", JOptionPane.WARNING_MESSAGE);
-        } else {
+        
+        if (Carpeta.exists() && Carpeta.isDirectory()) {
+            VaciarCarpeta(Carpeta);
             Carpeta.delete();
-            JOptionPane.showMessageDialog(null, "Carpeta eliminada de forma exitosa", null, JOptionPane.INFORMATION_MESSAGE);
+            JOptionPane.showMessageDialog(null, "Carpeta eliminada de forma exitosa", null, JOptionPane.INFORMATION_MESSAGE);////////////////////////
+        } else {
+            JOptionPane.showMessageDialog(null, "Carpeta no encontrada", "Error", JOptionPane.WARNING_MESSAGE);
         }
     }
 }
